requirements.txt
python-dotenv==1.1.0
openai==1.78.1
streamlit==1.45.1
PyPDF2==3.0.1
reportlab==4.4.1
streamlit-pdf-viewer==0.0.23
